DBZ-2534 Synchronize on engine start and stop
Current solution when we synchronize on `runningThread` is not
sufficient as e.g. sequence of start/stop/start in different threads
can result in state when the first thread runs `start()`, second calls
`stop()` and sets `runningThread` to `null`, but before the first thread
notices that `runningThread` is `null`, another thread calls `start()`
and as `runningThread` is `null` by that time, it will continue with
`start()` and therefore there will be two threads running `start()`
method on the same engine instance.

To prevent it, synchronize threads when setting `runningThread` and
also block during the whole `stop()` method to make sure no other thread
will call `start()` until stop finishes.

This doens't fix all possible issues, there is still possibility that
`stop()` will be called before `run()` is actually executed, but this
has to be coordinated by the upper level.