DBZ-5552 Fix randomly failing SqlServerConnectorIT
SqlServerConnectorIT.whenMultipleCaptureInstancesExcludesColumnsExpectLatestCDCTableUtilized
started to fail randomly in GitHub CI recently. It seems to be a timing
issue. Make test more predictable and read changes one by one, i.e.
check first snapshot events and after that change the schema, insert
new item and check the records.
The merit of the test should be preserved as it seems that the main
goal of the test is to test that exclude columns are not included in
the records.