DBZ-4490 Fix randomly failing updatePrimaryKeyWithRestartInMiddle test
Fix randomly failing
SqlServerConnectorIT#updatePrimaryKeyWithRestartInMiddle.

It seems the main source of instability in slow environments is
consuming snapshot record. It's actually not consumed and as the
test assumes exact order of messages, it leads to failures.

Wait for snapshot to be completed using mbean server. Also add more
asserts to make clear where is the issue if the test continues to fail.