DBZ-6035 Add function for adjusting SQL server polling interval
Some tests fail randomly as they insert some records at the begging
of the test, start Debezium and asserts that there are no records.
However, as SQL server poll for the changes and copies them into
CDC table not immediately but in specified interval (by default 5
secods), there can be a race condition when the entries inserted
at the beginning of the test are copied into CDC table after starting
Debezium and it resutls into the test failure as there are unexpected
records.

Add function for adjusting the polling interval and decrease polling
interval in `IncrementalSnapshotIT` which fails very often.

This doesn't fully fix the race condition, but decreases the changes
the issue is hit.