DBZ-5201: Add MongoDB connector support for server-side filtering of replica set change streams.
This change migrates from doing client-side filtering of streaming
change events to performing it server-side via MongoDB's support
for aggregation pipelines. This can drastically cut down on the amount
of data sent between the client and server. In addition, a new approach
to heartbeating has been added which relies on the `postBatchResumeToken`
to advance the the cursor even when no changes occur on the pipeline.