DBZ-6155 Adjust LogMiner batch size based on the default batch size
In DBZ-5005 the LogMiner batch size adjustment was changed to be based
on current batch size. DBZ-5005 as well as the commit itself is missing
any justification for it and it seems this change is actually wrong as

    topScnToMine = startScn + currentBatchSizeScn

so the condition

    topScnToMine - currentScn > currentBatchSizeScn

is actually

    startScn - currentScn > 0

i.e. that `startSCN` is before `currentScn` which very likely never
happens.

Revert is back to adjustment based `defaultBatchSize`. This is already
battle-tested value and also make the window and sleeping time more
stable than `logMiningBatchSizeMin` suggested in DBZ-6155. Using
`logMiningBatchSizeMin` would also result in very small window which
would mean we will send many small queries instead of several bigger
ones during mining.

Using current batch size for comparison is wrong in case that when
`currentScn` is `topScn` as shown above and in the other case (when
`topScn` is behind `currentScn`) is can eventually lead to situation
when `topScn` would fall behind `currentScn` more and more as we would
compare `currentScn - topScn` with bigger and bigger number (current
batch size).

Sketch of the two condition is now as follows:

* `topScn` is in "far future":

                                      <--default batch size-->
               <------------------------------batch size--------------------------->
    ---------|-----------------------|------------------------|---------------------|------
        startScn                 currentScn                                         topScn

* `currentScn` is more than `defaultBatchSize` away from `topScn`:

               <------------------batch size------------------------><-----default batch size------>
    ---------|-------------------------------------------------------|------------------------------|-----|------
        startScn                                                    topScn                               currentScn