Keep timezone information
Instead of converting everything to UTC, which effectively drop the
timezone information, continue to use the timezone which is read from
the socket.

pgjdbc will set the timezone of the connection to the VMs timezone by
default, therefore this has the negative side effect that the output
result depends on the VMs timezone.