DBZ-6725 Include beforeRecord if the delete handling mode is REWRITE
So when performing delete events, like other Event Flattening in other Relational databases shows the records before deletion for delete.handling.mode REWRITE. But, in case of MongoDB, the information of the record is not included in the delete events. This commit allows you to add the information of the record before it's deleted, along with __deleted is true.