DBZ-6176 Include records with lower bound SCN into result set
Several records may share same SCN and when we exclude records with
lower bound SCN from the result set, we may miss some records upon
restart. Therefore, we should include also records with lower bound
SCN. To avoid duplicates, we need to change upper bound for the query
not to be inclusive.