DBZ-2913 Fix some time precision failures
MySQL rounds precision up to match a column's precision when the precision is
smaller than the time value whereas MariaDB truncates by default. This can be
influenced on the MariaDB side by using `TIME_ROUND_FRACTIONAL` to avoid the
truncation.

In this commit, I've aligned the tests to MariaDB's default behavior.