Multi-task connector: implemented task assignment and event distribution (#63)
### Notify
cc stripe-internal/online-db-cdc
r? anthonyn shichao


### Summary

This PR adds the implementation of task assignment and event distribution for the multi-task connector setup. The changes touch the components in the diagram enclosed in red. The following features are now supported:
- Multiple tasks can be assigned to consume a single replset concurrently. The number of tasks is configured with `tasks.max`
- The feature can be turned on and off based on the config `mongodb.multi.task.enabled`
- Events are assigned to a task based on its timestamp. The current logic is the same as the multi-connector assignment logic: the task id is equal to `timestamp % numTasks`
  - A future work will include supporting event assignment based on other properties such as object id
- With this PR, we can spawn up brand new connectors with multiple tasks. But it currently does not work when the connector restarts without the offset management change.
<img width="942" alt="Screenshot 2023-12-02 at 5 53 51 PM" src="https://git.corp.stripe.com/storage/user/11372/files/7fb3c180-6c34-4abf-b7b4-274d4094e228">


### Motivation
https://jira.corp.stripe.com/browse/CDC-1424

### Test plan

- [x] Unit tests added to verify the correctness of task distributions
- [x] Integration tests added to verify the correctness of event distributions
  - Because the IT framework does not support having more than one running tasks, the current tests only verify correctness with `task 0`. In future when more than one task can be spawned we can update the testing logic to verify all tasks.
- [x] Tested with debezium-local-dev to spawn up 3 tasks and verified that the events are assigned based on task IDs and that the tasks only process the events that matched their IDs.

I set up the local debezium with this change and with `tasks.max=3`, and then changed the collection. For example, in the screenshot I performed delete, update, and then insert. In each case, we can see that two of the tasks skipped the event, and the one task processed the event, matching my expectation.
<img width="1080" alt="image" src="https://git.corp.stripe.com/storage/user/11372/files/d3e0d798-9b9e-4467-bb6b-c0826b9eda41">
<img width="1081" alt="image" src="https://git.corp.stripe.com/storage/user/11372/files/db463fd9-58b1-4e7b-aaf7-71fd63c19533">
<img width="1083" alt="image" src="https://git.corp.stripe.com/storage/user/11372/files/880db394-7ec3-4197-beb8-d0afa43f8833">


### Rollout/revert plan

Will roll out the change to debezium-mongo-srv afterward when the offset management is also checked in, by following the runbook [UpdateConnectorDependencies](https://confluence.corp.stripe.com/display/CDC/Operation%3A+Update+Connector+Dependencies#Operation:UpdateConnectorDependencies-debezium-connector-mongodb). We will ensure that for the time being, the feature is disabled via the config for all connectors except for the ones we are testing.

(Squashed by Merge Queue - Original PR: https://git.corp.stripe.com/stripe-private-oss-forks/debezium/pull/63)