DBZ-7368: add more documentation for converters in outbox router
Re-structure and add more documentation about setting up the converters in the Outbox Event Router SMT. My objective is make it clear how to properly use JSON and Avro regarding the value converters, the db column types, and the delegate converter configuration.

I'm also adding an extra example of how to setup a delegate converter in Avro format.

Needless to say, this is just a suggestion and is open to any comment or recommendation.

Jira [DBZ-7368](https://issues.redhat.com/browse/DBZ-7368)