DBZ-7024 Limit size of records processing thread pool
If the number of threads is provided as a number, limit it to 16 threads
to avoid possible overhead with context switching on a beefy machines
where the default value using all available cores may result in many
threads, which would be waiting most of the time anyway, as such machine
may run probably many other tasks, not only Debezium.

If the user really wants to use all available cores, it can be specified
using `AVAILABLE-CORES` placeholder.