[ci] Fix Oracle test workflow worker space failure
A GitHub worker only has so much space and currently when this stage of the job runs concurrently with the
"test-connector" stage, we are facing a worker out of space problem. To attempt to combat this issue, the
REST extension tests will run sequentially after the "test-connector" stage has concluded.