Stepwise support of multi-task for CES (#79)
### Notify
cc stripe-internal/online-db-cdc

### Summary
**MongoDB Multi-Task offset stepwise mechanism via hop.**
With stepwise mechanism we track the start / stop time to manage the cursor in change stream.

Each task is managed deterministically based on taskId (zero indexed), taskCount, and hop size.
**This allows DBZ to managed tasks separately and ensure that each event is only processed by one task.**

Example: 3 tasks, hop size of 5, and lastOffset of 0, will distribute the tasks based on offsetTime showing start and stop times.
- Task 0: [0->4], [15->19], [30->34], ...
- Task 1: [5->9], [20->24], ...
- Task 2: [10->14], [25->29], ...

**Beyond this we have the ability to optimize the start time.**
If the offset is in the middle of a step it will process from that point, reducing the hop size.
If the offset is past the stop time, it will move the next hopt

Example: 3 tasks, hop size of 5, and lastOffset of 7, will distribute the tasks based on offsetTime showing start and stop times.

- Task 0: [15->19], [30->34], ... (Skipping [0-4])
- Task 1: [7->9], [20->24], ...   (Reducing the first hop from [5->9] to [7->9]
- Task 2: [10->14], [25->29], ... (No changes)

Visual of logic
![image](https://git.corp.stripe.com/storage/user/14814/files/6d5f3715-172e-4f7e-a0bf-cd9ca89c92cf)

**Note:**
There are sections of the integration tests that have been commented out. This is due to the chance these will fail due to task distribution is based on natrual datetime, not when the test was started.
We will follow up to fix this via this JIRA
Update Multi-Task DBZ integration tests to be deterministic  - https://jira.corp.stripe.com/browse/CDC-1785

### Motivation
https://jira.corp.stripe.com/browse/CDC-1755


### Test plan
- [ ] All changes in PR are covered by tests
- [ ] Failures and edge cases tested
- [ ] Manually ran affected jobs

> UNIT TESTS
> $ mvn clean test -pl debezium-connector-mongodb
> [INFO] Results:
> [INFO]
> [INFO] Tests run: 184, Failures: 0, Errors: 0, Skipped: 0
> [INFO]
> [INFO] ------------------------------------------------------------------------
> [INFO] BUILD SUCCESS
> [INFO] ------------------------------------------------------------------------
> [INFO] Total time:  8.061 s
> [INFO] Finished at: 2024-04-15T15:42:19-07:00
> [INFO] ------------------------------------------------------------------------

> INTEGRATION TESTS
> $ mvn cean verify -DskipTests -pl debezium-connector-mongodb
> [INFO] Results:
> [INFO]
> [ERROR] Tests run: 257, Failures: 44, Errors: 1, Skipped: 0
> [INFO]
> [INFO] ------------------------------------------------------------------------
> [INFO] BUILD FAILURE
> [INFO] ------------------------------------------------------------------------
> [INFO] Total time:  10:50 min
> [INFO] Finished at: 2024-04-15T15:59:43-07:00
> [INFO] ------------------------------------------------------------------------

(Squashed by Merge Queue - Original PR: https://git.corp.stripe.com/stripe-private-oss-forks/debezium/pull/79)