[DBZ-PGYB] Use YugabyteDB smart driver instead of vanilla Postgres JDBC driver (#107)
## Problem

The Debezium connector for Postgres uses a single host model where the
JDBC driver connects to a PG instance and continues execution. However,
when we move to YugabyteDB where we have a multi node deployment, the
current model can fail in case the node it has connected to goes down.

## Solution

To address that, we have made changes in this PR and replaced the
Postgres JDBC driver with [YugabyteDB smart
driver](https://github.com/yugabyte/pgjdbc) which allows us to specify
multiple hosts in the JDBC url so that the connector does not fail or
run into any fatal error while maintaining the High Availability aspect
of YugabyteDB.

Changes in this PR include:
1. Changing of version in `pom.xml` from `2.5.2.Final` to
`2.5.2.ybpg.20241-SNAPSHOT`
a. This is done to ensure that upon image compilation, the changed code
from Debezium Code is picked up.
2. Replacing of all packages from `org.postgresql.*` to `com.yugabyte.*`
to comply with the new JDBC driver.
3. Masking the validator method in debezium-core which disallowed
characters like `: (colon)` in the configuration property
`database.hostname`