[DBZ-PGYB] Support for consistent snapshot for an existing slot (#113)
**Summary**
This PR is to support consistent snapshot in the case of an existing
slot.

In this case, the consistent_point hybrid time is determined from the
pg_replication_slots view, specifically from the yb_restart_commit_ht
column.

There is an assumption here that this slot has not been used for
streaming till this point. If this holds, then the history retention
barrier will be in place as of the consistent snapshot time
(consistent_point). The snapshot query will be run as of the
consistent_point and subsequent streaming will start from the
consistent_point of the slot.

**Test Plan**
Added new test
mvn -Dtest=PostgresConnectorIT#initialSnapshotWithExistingSlot test