Wait for snapshot completion acknowledgement before transitioning to streaming (#116)
## Problem

PG connector does not wait for acknowledgement of snapshot completion
offset before transitioning to streaming. This can lead to an issue if
there is a connector restart in the streaming phase and it goes for a
snapshot on restart. In streaming phase, as soon as the 1st GetChanges
call is made on the server, the retention barriers are lifted and so the
server can no longer serve the snapshot records on a restart. Therefore
it is important that the connector waits for acknowledgement of snapshot
completion offset before it actually transitions to streaming.

## Solution

This PR introduces a waiting mechanism for acknowledgement of snapshot
completion offset before transitioning to streaming.

We have introduced a custom heartbeat implementation that will dispatch
heartbeat when forced heartbeat method is called but we'll dispatch
nothing when a normal heartbeat method is called.

With this PR, connector will dispatch heartbeats while waiting for the
snapshot completion offset i.e during the transition phase. For these
heartbeat calls, there is no need to set the `heartbeat.interval.ms`
since we are making forced heartbeat calls which do not rely on this
config. Note, this heartbeat call is only required to support
applications using debezium engine/embedded engine. It is not required
when the connector is run with kakfa-connect.

### Test Plan
Manually deployed connector in a docker container and tested two
scenarios: 0 snapshot records & non-zero snapshot records. Unit tests
corresponding to these scenarios will be added in a separate PR.