[DBZ-PGYB] Changes to support all operations with all replica identities (#130)
This PR adds the changes to the connector required to support all kinds
of operations supported by YugabyteDB when coupled with various replica
identities.

One change to note here would be that in case of `DEFUALT` replica
identity, Postgres sends out the before image of the primary key column
in case of `DELETE` and `UPDATE` operations. Meanwhile, YugabyteDB only
sends out the before image of primary key in case of `DELETE`
operations, and if a primary key itself is updated, YugabyteDB sends two
events:
1. Delete for the existing row
2. Insert with the new value of the primary key

Additionally to test out everything, we have a new test class
`YugabyteReplicaIdentityIT`.